chore: migrate BenisInNavbar module
